Create GuessMasterNumber.bat
When the game starts, the code will generate a random number between 1 and 100 and store it in the `answer` variable. It will then prompt the user to guess the number.

If the user's guess is correct, the game will congratulate them and display the number of attempts it took to guess the correct answer. If their guess is too low, the game will prompt the user to guess again and inform them that their guess was too low. If their guess is too high, the game will prompt the user to guess again and inform them that their guess was too high.